Use PlayerProx component to trigger inventory refresh
When a player is near a chest, the player's inventory will be refreshed, adding the items in the chest to the number of total available resources they can use to craft items.
When the player is far from the chest, the number of available resources will go back down to normal.