Override GetCraftingIngredients
GetCraftingIngredients is a new function Klei added that we had to override because the game couldn't find items from closed nearby chests meaning they would not be consumed. The client code is unchanged because Klei didn't make major changes to the way ingredients are removed. Also, a parameter was added to a few Klei functions, so we accounted for that.